make plot and data for 3 exercise

+++ GTest/test_for_csr_matrix.cpp
@@ -15,7 +15,7 @@ TEST(CSRMatrix, Constructor) {
                       {{3, 3}, 11.1},
                       {{3, 6}, 6.0006}
     }, 4, 7);
-    vector<float>  values = {1.0, 19.1, 99.5, 1.98, 2.009, 3.11, 9.999, 8.0, 11.1, 6.0006};
+    vector<double>  values = {1.0, 19.1, 99.5, 1.98, 2.009, 3.11, 9.999, 8.0, 11.1, 6.0006};
     vector<size_t> columns_indexes = {0, 2, 5, 1, 2, 3, 4, 2, 3, 6};
     vector<int> number_rows_non_0_elements = {0, 3, 6, 7, 10};
 
@@ -51,38 +51,38 @@ TEST(CSRMatrix, Multiplication) {
         {{3, 3}, 11.1},
         {{3, 6}, 6.0006}
     }, 4, 7);
-    vector<float> column1_for_multiplication = {1.0, 2.99, 3.0001, 11.21, 31.1, 26.878, 4.0};
-    vector<float> check_answer1 = {2732.66291, 46.8105009, 310.9689, 172.4342};
-    vector<float> column2_for_multiplication = {0, 0, 0, 1, 0, 0, 0};
-    vector<float> check_answer2 = {0, 3.11, 0, 11.1};
-    vector<float> column3_for_multiplication = {0, 0, 1, 0, 0, 0, 0};
-    vector<float> check_answer3 = {19.1, 2.009, 0, 8};
-    vector<float> column4_for_multiplication = {0, 1, 0, 0, 0, 0, 0};
-    vector<float> check_answer4 = {0, 1.98, 0, 0};
-    vector<float> column5_for_multiplication = {1, 0, 0, 0, 0, 0, 0};
-    vector<float> check_answer5 = {1, 0, 0, 0};
-    vector<float> column6_for_multiplication = {0, 0, 0, 0, 1, 0, 0};
-    vector<float> check_answer6 = {0, 0, 9.999, 0};
-    vector<float> column7_for_multiplication = {0, 0, 0, 0, 0, 1, 0};
-    vector<float> check_answer7 = {99.5, 0, 0, 0};
-    vector<float> column8_for_multiplication = {0, 0, 0, 0, 0, 0, 1};
-    vector<float> check_answer8 = {0, 0, 0, 6.0006};
-    vector<float> column9_for_multiplication = {2.78, 1.56, 0, 23.88, 6.09, 45.765, 8.654};
-    vector<float> check_answer9 = {4556.3975, 77.3556, 60.89391, 316.9971924};
-    vector<float> zero_column_for_multiplication = {0, 0, 0, 0, 0, 0, 0};
-    vector<float> zero_check_answer = {0, 0, 0, 0};
+    vector<double> column1_for_multiplication = {1.0, 2.99, 3.0001, 11.21, 31.1, 26.878, 4.0};
+    vector<double> check_answer1 = {2732.66291, 46.8105009, 310.9689, 172.4342};
+    vector<double> column2_for_multiplication = {0, 0, 0, 1, 0, 0, 0};
+    vector<double> check_answer2 = {0, 3.11, 0, 11.1};
+    vector<double> column3_for_multiplication = {0, 0, 1, 0, 0, 0, 0};
+    vector<double> check_answer3 = {19.1, 2.009, 0, 8};
+    vector<double> column4_for_multiplication = {0, 1, 0, 0, 0, 0, 0};
+    vector<double> check_answer4 = {0, 1.98, 0, 0};
+    vector<double> column5_for_multiplication = {1, 0, 0, 0, 0, 0, 0};
+    vector<double> check_answer5 = {1, 0, 0, 0};
+    vector<double> column6_for_multiplication = {0, 0, 0, 0, 1, 0, 0};
+    vector<double> check_answer6 = {0, 0, 9.999, 0};
+    vector<double> column7_for_multiplication = {0, 0, 0, 0, 0, 1, 0};
+    vector<double> check_answer7 = {99.5, 0, 0, 0};
+    vector<double> column8_for_multiplication = {0, 0, 0, 0, 0, 0, 1};
+    vector<double> check_answer8 = {0, 0, 0, 6.0006};
+    vector<double> column9_for_multiplication = {2.78, 1.56, 0, 23.88, 6.09, 45.765, 8.654};
+    vector<double> check_answer9 = {4556.3975, 77.3556, 60.89391, 316.9971924};
+    vector<double> zero_column_for_multiplication = {0, 0, 0, 0, 0, 0, 0};
+    vector<double> zero_check_answer = {0, 0, 0, 0};
 
 
-    vector<float> answer1 = matrix * column1_for_multiplication;
-    vector<float> answer2 = matrix * column2_for_multiplication;
-    vector<float> answer3 = matrix * column3_for_multiplication;
-    vector<float> answer4 = matrix * column4_for_multiplication;
-    vector<float> answer5 = matrix * column5_for_multiplication;
-    vector<float> answer6 = matrix * column6_for_multiplication;
-    vector<float> answer7 = matrix * column7_for_multiplication;
-    vector<float> answer8 = matrix * column8_for_multiplication;
-    vector<float> answer9 = matrix * column9_for_multiplication;
-    vector<float> zero_answer = matrix * zero_column_for_multiplication;
+    vector<double> answer1 = matrix * column1_for_multiplication;
+    vector<double> answer2 = matrix * column2_for_multiplication;
+    vector<double> answer3 = matrix * column3_for_multiplication;
+    vector<double> answer4 = matrix * column4_for_multiplication;
+    vector<double> answer5 = matrix * column5_for_multiplication;
+    vector<double> answer6 = matrix * column6_for_multiplication;
+    vector<double> answer7 = matrix * column7_for_multiplication;
+    vector<double> answer8 = matrix * column8_for_multiplication;
+    vector<double> answer9 = matrix * column9_for_multiplication;
+    vector<double> zero_answer = matrix * zero_column_for_multiplication;
 
 
     for(auto number_element = 0; number_element < answer1.size(); ++ number_element){
