+++ solving/QRDecomposition/householder.cpp
@@ -2,17 +2,17 @@
 
 #import <cmath>
 
-vector<float> MakeBasisVector(const int &number_elements) {
-    vector<float> basis_vector(0.0, number_elements);
+vector<double> MakeBasisVector(const int &number_elements) {
+    vector<double> basis_vector(0.0, number_elements);
     basis_vector[0] = 1.0;
     return basis_vector;
 }
 
-vector<float> FindNormal(const vector<float> &matrix_column) {
-    vector<float> basis_vector = MakeBasisVector(matrix_column.size());
+vector<double> FindNormal(const vector<double> &matrix_column) {
+    vector<double> basis_vector = MakeBasisVector(matrix_column.size());
     bool first_coordinate_more_null = matrix_column[0] > 0.0;
-    vector<float> matrix_column_normal;
-    float matrix_column_length = sqrt(matrix_column * matrix_column);
+    vector<double> matrix_column_normal;
+    double matrix_column_length = sqrt(matrix_column * matrix_column);
     if (first_coordinate_more_null) {
         matrix_column_normal = matrix_column + matrix_column_length * basis_vector;
     } else {
@@ -21,8 +21,8 @@ vector<float> FindNormal(const vector<float> &matrix_column) {
     return matrix_column_normal;
 }
 
-Matrix MultiplyColumnRow(const vector<float> &column, const vector<float> &row) {
-    vector<float> matrix;
+Matrix MultiplyColumnRow(const vector<double> &column, const vector<double> &row) {
+    vector<double> matrix;
     matrix.reserve(row.size() * column.size());
     for (int number_column_element = 0; number_column_element < column.size();
          ++number_column_element) {
@@ -35,9 +35,9 @@ Matrix MultiplyColumnRow(const vector<float> &column, const vector<float> &row) 
     return {matrix, row.size()};
 }
 
-Matrix MakeTransformationMatrix(const vector<float> &normal) {
+Matrix MakeTransformationMatrix(const vector<double> &normal) {
     int size_transformation_matrix = normal.size();
-    vector<float> unit_matrix(0, size_transformation_matrix);
+    vector<double> unit_matrix(0, size_transformation_matrix);
     for (int number_row = 0; number_row < size_transformation_matrix; ++number_row) {
         unit_matrix[number_row * size_transformation_matrix + number_row] = 1;
     }
@@ -47,18 +47,18 @@ Matrix MakeTransformationMatrix(const vector<float> &normal) {
     return transformation_matrix;
 }
 
-vector<float> ConvertsMatrixColumnOrRow(const vector<float> &matrix_column,
-                                       const vector<float> &normal, const float &denominator) {
-    float numerator = matrix_column * normal;
-    vector<float> converted_column = matrix_column - 2.0 * numerator / denominator * normal;
+vector<double> ConvertsMatrixColumnOrRow(const vector<double> &matrix_column,
+                                       const vector<double> &normal, const double &denominator) {
+    double numerator = matrix_column * normal;
+    vector<double> converted_column = matrix_column - 2.0 * numerator / denominator * normal;
     return converted_column;
 }
 
 void TransformUpperDiagonalMatrix(Matrix &matrix, const int &number_transformation,
-                                  const vector<float> &normal, const float &denominator) {
+                                  const vector<double> &normal, const double &denominator) {
     for (int number_column = number_transformation; number_column < matrix.GiveNumberColumns();
          ++number_column) {
-        vector<float> converted_column =
+        vector<double> converted_column =
             matrix.GiveNestedColumn(number_transformation, number_column);
         converted_column = ConvertsMatrixColumnOrRow(converted_column, normal, denominator);
         matrix.ModifyNestedColumn(converted_column, number_column);
@@ -66,9 +66,9 @@ void TransformUpperDiagonalMatrix(Matrix &matrix, const int &number_transformati
 }
 
 void TransformOrthogonalMatrix(Matrix &matrix, const int &number_transformation,
-                               const vector<float> &normal, const float &denominator) {
+                               const vector<double> &normal, const double &denominator) {
     for (int number_row = 0; number_row < matrix.GiveNumberColumns(); ++number_row) {
-        vector<float> converted_row = matrix.GiveNestedRow(0, number_row);
+        vector<double> converted_row = matrix.GiveNestedRow(0, number_row);
         converted_row = ConvertsMatrixColumnOrRow(converted_row, normal, denominator);
         matrix.ModifyNestedRow(converted_row, number_row);
     }
@@ -76,15 +76,15 @@ void TransformOrthogonalMatrix(Matrix &matrix, const int &number_transformation,
 
 QRDecomposition HouseholderAlgorithm(const Matrix &original_matrix) {
     Matrix upper_diagonal_matrix = original_matrix;
-    vector<float> matrix_column = {upper_diagonal_matrix.GiveNestedColumn(0, 0)};
-    vector<float> normal = FindNormal(matrix_column);
+    vector<double> matrix_column = {upper_diagonal_matrix.GiveNestedColumn(0, 0)};
+    vector<double> normal = FindNormal(matrix_column);
     Matrix transformation_matrix = MakeTransformationMatrix(normal);
     for (int number_transformation = 0; number_transformation < original_matrix.GiveNumberColumns();
          ++number_transformation) {
         matrix_column = {
             upper_diagonal_matrix.GiveNestedColumn(number_transformation, number_transformation)};
         normal = FindNormal(matrix_column);
-        float denominator = normal * normal;
+        double denominator = normal * normal;
         TransformUpperDiagonalMatrix(upper_diagonal_matrix, number_transformation, normal,
                                      denominator);
         TransformOrthogonalMatrix(transformation_matrix, number_transformation, normal,
